Changed return type of PlayerInventory::getItemAmountBySlot to short as it should be.
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@690 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerInventory.h
@@ -144,7 +144,7 @@ public:
 		else
 			items[inv][slot] = item;
 	}
-	int getItemAmountBySlot(char inv, short slot) {
+	short getItemAmountBySlot(char inv, short slot) {
 		inv -= 2;
 		if (items[inv].find(slot) != items[inv].end())
 			return items[inv][slot]->amount;
